update imputation figure.
