+++ msresist/expectation_maximization.py
@@ -46,7 +46,7 @@ def EM_clustering(data, info, ncl, seqDist=None, gmmIn=None):
         else:
             gmm = gmmIn
 
-        gmm.fit(d, max_iterations=500, verbose=True, stop_threshold=1e-6)
+        gmm.fit(d, max_iterations=500, verbose=True, stop_threshold=1e-4)
         scores = gmm.predict_proba(d)
 
         if np.all(np.isfinite(scores)):
